dashboard/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
anthropic>=0.40
pydantic>=2.0
websockets>=12.0
